Add handling page slugs
