+++ app/(main)/pages/[idOrSlug]/page.module.scss
